feat: render ExitPlanMode output as markdown and remove tool grouping
Render each tool call as an individual collapsible accordion instead of
grouping consecutive tool calls into a single nested accordion. ExitPlanMode
tool output is rendered as formatted markdown below its accordion.

Closes #40
