+++ apps/web/src/components/ai-elements/group-parts.ts
@@ -11,51 +11,33 @@ export type TextSegment = {
   partIndex: number;
 };
 
-export type ToolGroupSegment = {
-  type: "tool-group";
-  parts: Array<{ part: ToolPart; partIndex: number }>;
-  /** Index of the first tool part in the group, used as a stable React key */
-  startIndex: number;
+export type ToolSegment = {
+  type: "tool";
+  part: ToolPart;
+  partIndex: number;
 };
 
-export type MessageSegment = TextSegment | ToolGroupSegment;
+export type MessageSegment = TextSegment | ToolSegment;
 
 /**
- * Groups consecutive tool UI parts into ToolGroupSegments.
- * Empty/whitespace-only text parts are skipped and do not break tool grouping.
- * All other non-tool parts become individual TextSegments.
+ * Separates message parts into text and tool segments.
+ * Empty/whitespace-only text parts are skipped.
  */
 export function groupMessageParts(parts: Part[]): MessageSegment[] {
   const segments: MessageSegment[] = [];
-  let currentToolGroup: ToolGroupSegment | null = null;
-
-  const flushToolGroup = () => {
-    if (currentToolGroup) {
-      segments.push(currentToolGroup);
-      currentToolGroup = null;
-    }
-  };
 
   for (let i = 0; i < parts.length; i++) {
     const part = parts[i];
 
     if (isToolUIPart(part)) {
-      if (!currentToolGroup) {
-        currentToolGroup = { type: "tool-group", parts: [], startIndex: i };
-      }
-      currentToolGroup.parts.push({ part: part as ToolPart, partIndex: i });
+      segments.push({ type: "tool", part: part as ToolPart, partIndex: i });
       continue;
     }
 
-    // Only non-empty text parts break tool grouping and become segments.
-    // Everything else (empty text, step-start, reasoning, source-url, etc.)
-    // is skipped so consecutive tool calls stay grouped.
     if (part.type === "text" && part.text.trim()) {
-      flushToolGroup();
       segments.push({ type: "text", part, partIndex: i });
     }
   }
 
-  flushToolGroup();
   return segments;
 }
